model_monitoring_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiohttp==3.11.11
asyncio==3.4.3
python-dotenv==1.0.1    
scikit-learn==1.6.1
SQLAlchemy==2.0.36